// src/types/renshuuApiTypes.ts
export interface paths {
// ...
    "/lists": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List of all lists made by user, grouped by termtype and groups */
        get: {
            parameters: {
                query?: never;
                header?: never;
                path?: never;
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            termtype_groups?: {
                                /** @enum {string} */
                                termtype?:
                                    | "vocab"
                                    | "kanji"
                                    | "grammar"
                                    | "sent";
                                list_count?: number;
                                groups?: {
                                    list_count?: number;
                                    group_title?: string;
                                    lists?: components["schemas"]["List"][];
                                }[];
                            }[];
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/list/{id}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List of all terms in the specified user list */
        get: {
            parameters: {
                query?: {
                    /** @description Page of results */
                    pg?: string;
                };
                header?: never;
                path: {
                    /** @description The list id */
                    id: string;
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": components["schemas"]["List"];
                    };
                };
                /** @description List could not be found. */
                404: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content?: never;
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
    "/list/all/{termtype}": {
        parameters: {
            query?: never;
            header?: never;
            path?: never;
            cookie?: never;
        };
        /** List of all terms of that type studied by the user */
        get: {
            parameters: {
                query?: {
                    /** @description Page of results */
                    pg?: string;
                };
                header?: never;
                path: {
                    /** @description The type of terms */
                    termtype: "vocab" | "kanji" | "grammar" | "sent";
                };
                cookie?: never;
            };
            requestBody?: never;
            responses: {
                /** @description Successful response */
                200: {
                    headers: {
                        [name: string]: unknown;
                    };
                    content: {
                        "application/json": {
                            contents?: {
                                /** @example 1 */
                                result_count?: number;
                                /** @example 2 */
                                total_pg?: number;
                                /** @example 50 */
                                per_pg?: number;
                                /** @example 1 */
                                pg?: number;
                                /** @example all */
                                group?: string;
                                terms?: (
                                    | components["schemas"]["Word"]
                                    | components["schemas"]["Kanji"]
                                    | components["schemas"]["Grammar"]
                                    | components["schemas"]["SimpleSentence"]
                                )[];
                            };
                        };
                    };
                };
            };
        };
        put?: never;
        post?: never;
        delete?: never;
        options?: never;
        head?: never;
        patch?: never;
        trace?: never;
    };
}
export type webhooks = Record<string, never>;
export interface components {
// ...
    responses: never;
    parameters: never;
    requestBodies: never;
    headers: never;
    pathItems: never;
}